Restructured packages in Rent-Main.
