+++ Rent/Rent-Main/src/main/java/Application.java
@@ -1,5 +1,3 @@
-package Wang;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.ApplicationContext;
@@ -9,7 +7,7 @@ import org.springframework.context.annotation.Configuration;
 import java.util.Arrays;
 
 @Configuration
-@ComponentScan("Configuration, Wang")
+@ComponentScan("Configuration, Controllers")
 @EnableAutoConfiguration
 public class Application {
